requirements.txt
pandas
pytest
streamlit
sentence-transformers
numpy
openai
python-dotenv
